// src/domain/interfaces/guide-transport.ts
export interface GuideTransport {
  id?: number;
  fecha: Date;
  fecha_sistema: Date;
  serie: string;
  numero: string;
  id_empresa: number;
  fl_estado: number;
  motivo_anulacion?: string;
  id_local: number;
  id_usuario: number;
  fecha_traslado?: Date;
  fl_guia_interna?: number;
  id_programacion_viaje?: number;
  nota?: string;
  respuesta_api?: string;
  estado?: string;
  json_api?: string;
  id_motivo_traslado?: number;
  id_orden?: number;
  fl_finalizado?: number;
  id_desembarque?: number;
  fl_multiple_viaje?: number;
  id_reparto?: bigint;
  enlace?: string;
  estadoSunat?: string;
  sunatDescription?: string;
  sunatNote?: string;
  sunatResponseCode?: string;
  sunatSoapError?: string;
  codigoQr?: string;
  nubefactError?: string;
  isElectronic?: string;
  pesoTotal?: number;
  id_remitente?: number;
  id_destinatario?: number;
  id_ubigeo_partida?: string;
  id_ubigeo_llegada?: string;
  punto_partida?: string;
  punto_llegada?: string;
  documento_relacionado_guiaremitente?: string;
  sunat_envio_indicador?: string;
  idSubcontratista?: number;
  idPagadorFlete?: number;
  pesoBrutoUnidadMedida?: string;
  nubefactResponseCode?: string;
  nubefactNotaImportante?: string;
}